+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Change Log
 
-All notable changes to the "requestinterceptmocking" extension will be documented in this file.
+All notable changes to the **proxyMocker** extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.1]
 
-- Initial release
+### Added
+
+- Proxy server
+- Save HTTP request mocks
+- Replace real HTTP request by mocks
+- Manage mocks